moved homepage-map, adding to readme
