+++ src/components/Interactive_map_comps/Homepage-map/Marks.js
@@ -2,9 +2,9 @@ import { geoIdentity, geoPath } from "d3";
 import { DataFilter } from "./DataFilter";
 import React, { useEffect, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import allData from "../content/json_choropleth/allData.json";
-import stateNameToAbbreviation from "../components/Abbrevinator";
-import "../components/styles/Marks.css";
+import allData from "../../../content/json_choropleth/allData.json";
+import stateNameToAbbreviation from "../../Abbrevinator.js";
+import "../../styles/Marks.css";
 
 const projection = geoIdentity().reflectY(false);
 const path = geoPath(projection);
